Don't create branches from branches created by this script.
This script finds branches from the Subversion branches, and then it creates
local references to those branches in the repository.
However, if those branches has been pushed to a remote server, Git creates
references to those remote references in 'origin/FOO', which were then retracked
by my sript, which would have then retracked those new branches, and so on.

So, as a shortcut, it now doesn't track the references pushed to an 'origin'
remote anymore. It's a bit dirty, but it does the job.